Move the filetestcase out of the code directory.
