+++ tests/test_gjslint.py
@@ -28,7 +28,7 @@ import unittest
 from jscodestyle import error_check
 from jscodestyle import errors
 from jscodestyle import runner
-from jscodestyle.tools import filetestcase
+from testtools import filetestcase
 
 
 _RESOURCE_PREFIX = 'tests/testdata'
