+++ tests/test_notstrict.py
@@ -26,7 +26,7 @@ import unittest
 
 from jscodestyle import errors
 from jscodestyle import runner
-from jscodestyle.tools import filetestcase
+from testtools import filetestcase
 
 _RESOURCE_PREFIX = 'tests/testdata'
